Fix circular deps for Cosmos (#2076)
