+++ src/cow-react/modules/wallet/api/components/Web3Status/Web3StatusMod.tsx
@@ -2,26 +2,21 @@
 import { t, Trans } from '@lingui/macro'
 import { useWeb3React } from '@web3-react/core'
 import { getConnection } from '@cow/modules/wallet/api/utils'
-import { darken } from 'polished'
 // import { useMemo } from 'react'
-import { Activity } from 'react-feather'
 import { useAppSelector } from 'state/hooks'
-import styled, { css } from 'styled-components/macro'
 
 import { useHasSocks } from 'hooks/useSocksBalance'
 import { useToggleWalletModal } from 'state/application/hooks'
 // import { isTransactionRecent, useAllTransactions } from 'state/transactions/hooks'
 // import { TransactionDetails } from 'state/transactions/types'
 import { shortenAddress } from 'utils'
-// import { ButtonSecondary } from 'components/Button'
 import StatusIcon from '@cow/modules/wallet/api/components/StatusIcon'
 import Loader from 'components/Loader'
 import { RowBetween } from 'components/Row'
 // import WalletModal from '../WalletModal'
-
 // MOD imports
 import FollowPendingTxPopup, { useCloseFollowTxPopupIfNotPendingOrder } from 'components/Popups/FollowPendingTxPopup'
-import { ButtonSecondary } from 'components/Button'
+import { NetworkIcon, Text, Web3StatusConnect, Web3StatusConnected, Web3StatusError } from './styled'
 
 /* const IconWrapper = styled.div<{ size?: number }>`
   ${({ theme }) => theme.flexColumnNoWrap};
@@ -32,83 +27,6 @@ import { ButtonSecondary } from 'components/Button'
     width: ${({ size }) => (size ? size + 'px' : '32px')};
   }
 `*/
-
-// mod
-export const Web3StatusGeneric = styled(ButtonSecondary)``
-
-const Web3StatusError = styled(Web3StatusGeneric)`
-  background-color: ${({ theme }) => theme.red1};
-  border: 1px solid ${({ theme }) => theme.red1};
-  color: ${({ theme }) => theme.white};
-  font-weight: 500;
-  :hover,
-  :focus {
-    background-color: ${({ theme }) => darken(0.1, theme.red1)};
-  }
-`
-
-export const Web3StatusConnect = styled(Web3StatusGeneric)<{ faded?: boolean }>`
-  /* border: none; */
-  /* color: ${({ theme }) => theme.primaryText1}; */
-  /* font-weight: 500; */
-
-  /* :hover,
-  :focus {
-    border: 1px solid ${({ theme }) => darken(0.05, theme.primary4)};
-    color: ${({ theme }) => theme.primaryText1};
-  } */
-
-  /* ${({ faded }) =>
-    faded &&
-    css`
-      background-color: ${({ theme }) => theme.primary5};
-      border: 1px solid ${({ theme }) => theme.primary5};
-      color: ${({ theme }) => theme.primaryText1};
-
-      :hover,
-      :focus {
-        border: 1px solid ${({ theme }) => darken(0.05, theme.primary4)};
-        color: ${({ theme }) => darken(0.05, theme.primaryText1)};
-      }
-    `} */
-`
-
-export const Web3StatusConnected = styled(Web3StatusGeneric)<{ pending?: boolean; clickDisabled?: boolean }>`
-  background-color: ${({ theme }) => theme.grey1};
-  border: 1px solid transparent;
-  color: ${({ theme }) => theme.text1};
-  font-weight: 500;
-
-  &:hover {
-    background-color: ${({ theme }) => theme.grey1};
-    color: ${({ theme }) => theme.text1};
-  }
-
-  ${({ clickDisabled }) =>
-    clickDisabled &&
-    css`
-      cursor: not-allowed;
-      pointer-events: none;
-    `}
-`
-
-export const Text = styled.p`
-  flex: 1 1 auto;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  margin: 0 0.5rem 0 0.25rem;
-  font-size: 1rem;
-  width: fit-content;
-  font-weight: 500;
-`
-
-const NetworkIcon = styled(Activity)`
-  margin-left: 0.25rem;
-  margin-right: 0.5rem;
-  width: 16px;
-  height: 16px;
-`
 
 // we want the latest one to come first, so return negative if a is after b
 // function newTransactionsFirst(a: TransactionDetails, b: TransactionDetails) {
