+++ src/cow-react/modules/wallet/api/components/Web3Status/index.tsx
@@ -1,8 +1,9 @@
 import styled from 'styled-components/macro'
-import { Web3StatusInner, Web3StatusConnected } from './Web3StatusMod'
+import { Web3StatusInner } from './Web3StatusMod'
 import { useWalletInfo, WalletModal } from '@cow/modules/wallet'
 import { STORAGE_KEY_LAST_PROVIDER } from 'constants/index'
 import { useCategorizeRecentActivity } from '@cow/common/hooks/useCategorizeRecentActivity'
+import { Web3StatusConnected } from './styled'
 
 export const Wrapper = styled.div`
   color: ${({ theme }) => theme.text1};
